refactor: move favorite button to separate file
