+++ src/components/products/product-card-base.tsx
@@ -1,47 +1,13 @@
-"use client";
-
 import { Product } from "@/api/product";
-import useFavorite from "@/hooks/use-favorite";
 import { cn } from "@/lib/util";
-import { HeartIcon } from "lucide-react";
-import { HTMLAttributes, MouseEvent, useState } from "react";
+import { HTMLAttributes } from "react";
 import { Badge } from "../ui/badge";
 import { Button } from "../ui/button";
 import ImageWithFallback from "../util/image-with-fallback";
+import ProductFavoriteButton from "./product-favorite";
 
 export interface ProductCardProps extends HTMLAttributes<HTMLDivElement> {
   item: Product;
-}
-
-function ProductCardFavorite({ item }: { item: Product }) {
-  const [isFavorite, setFavorite] = useFavorite(item.id);
-  const [heartBeat, setHeartbeat] = useState(false);
-
-  const elementClass =
-    "text-primary/30 transition hover:text-primary/50 hover:scale-110";
-
-  const clickHandler = (e: MouseEvent) => {
-    setHeartbeat(true);
-    setFavorite(!isFavorite);
-  };
-
-  return (
-    <div className="absolute right-4 top-4 z-10">
-      <Button
-        variant="text"
-        size="icon"
-        title="Добавить в избранное"
-        className={cn(
-          elementClass,
-          heartBeat && "animate-[heart-pulse_ease-in-out_0.5s]"
-        )}
-        onClick={clickHandler}
-        onAnimationEnd={(e) => setHeartbeat(false)}
-      >
-        <HeartIcon fill={isFavorite ? "hsl(var(--accent))" : "transparent"} />
-      </Button>
-    </div>
-  );
 }
 
 export default function ProductCard({
@@ -59,8 +25,12 @@ export default function ProductCard({
       )}
       {...props}
     >
-      <ProductCardFavorite item={item} />
-
+      <div className="absolute right-4 top-4 z-10">
+        <ProductFavoriteButton
+          productId={item!.id}
+          className="opacity-30 data-[active=true]:opacity-80"
+        />
+      </div>
       <div className="relative w-full aspect-square">
         <ImageWithFallback
           className="object-contain"
